requirements.txt
asgiref==3.2.7
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.6
django-filter==2.3.0
djangorestframework==3.11.0
fake-useragent==0.1.11
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
libsass==0.20.0
lxml==4.5.0
Markdown==3.2.2
mccabe==0.6.1
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.4.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.25.7
wrapt==1.11.2
